deprecated the class that starts with Hi
