+++ src/main/java/com/aliyun/hitsdb/client/callback/http/HttpResponseCallbackFactory.java
@@ -3,6 +3,7 @@ package com.aliyun.hitsdb.client.callback.http;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import com.aliyun.hitsdb.client.Config;
 import com.aliyun.hitsdb.client.util.Objects;
 import org.apache.http.HttpResponse;
 import org.apache.http.concurrent.FutureCallback;
@@ -19,7 +20,7 @@ public class HttpResponseCallbackFactory {
     private final HttpClient hitsdbHttpclient;
     private final boolean httpCompress;
 
-    public HttpResponseCallbackFactory(AtomicInteger unCompletedTaskNum,HttpClient httpclient,boolean httpCompress) {
+    public HttpResponseCallbackFactory(AtomicInteger unCompletedTaskNum, HttpClient httpclient, boolean httpCompress) {
         this.unCompletedTaskNum = unCompletedTaskNum;
         this.hitsdbHttpclient = httpclient;
         this.httpCompress = httpCompress;
@@ -31,45 +32,45 @@ public class HttpResponseCallbackFactory {
     }
 
     public FutureCallback<HttpResponse> createQueryCallback(final String address, final QueryCallback callback, final Query query) {
-        FutureCallback<HttpResponse> httpCallback = new QueryHttpResponseCallback(address, query, callback,this.httpCompress);
+        FutureCallback<HttpResponse> httpCallback = new QueryHttpResponseCallback(address, query, callback, this.httpCompress);
         return httpCallback;
     }
 
     public FutureCallback<HttpResponse> createBatchPutDataCallback(
-    			final String address,
+            final String address,
             final AbstractBatchPutCallback<?> batchPutCallback,
             final List<Point> pointList,
-            final HiTSDBConfig config,
+            final Config config,
             final int batchPutRetryCount
     ) {
-        FutureCallback<HttpResponse> httpCallback = new BatchPutHttpResponseCallback (
-					address,
-					hitsdbHttpclient,
-					batchPutCallback,
-					pointList,
-					config,
-					config.getBatchPutRetryCount()
-                );
+        FutureCallback<HttpResponse> httpCallback = new BatchPutHttpResponseCallback(
+                address,
+                hitsdbHttpclient,
+                batchPutCallback,
+                pointList,
+                config,
+                config.getBatchPutRetryCount()
+        );
         return httpCallback;
     }
-    
-    
+
+
     public FutureCallback<HttpResponse> createNoLogicBatchPutHttpFutureCallback(
-    			final String address,
+            final String address,
             final List<Point> pointList,
-            final HiTSDBConfig config,
+            final Config config,
             final int batchPutRetryTimes
     ) {
-        FutureCallback<HttpResponse> httpCallback = 
-                new BatchPutHttpResponseCallback (
-					address,
-					hitsdbHttpclient,
-					null,
-					pointList,
-					config,
-					batchPutRetryTimes
+        FutureCallback<HttpResponse> httpCallback =
+                new BatchPutHttpResponseCallback(
+                        address,
+                        hitsdbHttpclient,
+                        null,
+                        pointList,
+                        config,
+                        batchPutRetryTimes
                 );
         return httpCallback;
     }
-    
+
 }
