+++ src/main/java/com/aliyun/hitsdb/client/http/HttpAddressManager.java
@@ -1,19 +1,19 @@
 package com.aliyun.hitsdb.client.http;
 
-import com.aliyun.hitsdb.client.HiTSDBConfig;
+import com.aliyun.hitsdb.client.Config;
 
 public class HttpAddressManager {
 	private final String host;
 	private final int port;
 	private final String address;
 
-	private HttpAddressManager(HiTSDBConfig config) {
+	private HttpAddressManager(Config config) {
 		host = config.getHost();
 		port = config.getPort();
 		address = host + ":" + port;
 	}
 
-	public static HttpAddressManager createHttpAddressManager(HiTSDBConfig config) {
+	public static HttpAddressManager createHttpAddressManager(Config config) {
 		HttpAddressManager httpAddressManager = new HttpAddressManager(config);
 		return httpAddressManager;
 	}
