+++ src/main/java/com/aliyun/hitsdb/client/http/HttpClient.java
@@ -14,6 +14,7 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.zip.GZIPOutputStream;
 
+import com.aliyun.hitsdb.client.Config;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.http.Header;
 import org.apache.http.HttpHeaders;
@@ -103,7 +104,7 @@ public class HttpClient {
 		this.sslEnable = sslEnable;
 	}
 
-	HttpClient(HiTSDBConfig config, CloseableHttpAsyncClient httpclient, SemaphoreManager semaphoreManager, ScheduledExecutorService connectionGcService)
+	HttpClient(Config config, CloseableHttpAsyncClient httpclient, SemaphoreManager semaphoreManager, ScheduledExecutorService connectionGcService)
 			throws HttpClientInitException {
 		this.host = config.getHost();
 		this.port = config.getPort();
